--- src/presenter/routes/auth/auth_objects/mod.rs
@@ -1,3 +1,0 @@
-pub mod login_request;
-pub mod refresh_request;
-pub mod registration_request;
