--- src/presenter/routes/auth/auth_objects/refresh_request.rs
@@ -1,7 +1,0 @@
-use serde::Deserialize;
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct RefreshRequest<'a> {
-    #[serde(rename = "refresh_token")]
-    pub refresh_token: &'a str,
-}
