--- src/presenter/routes/auth/auth_objects/registration_request.rs
@@ -1,11 +1,0 @@
-use serde::Deserialize;
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct RegistrationRequest<'a> {
-    #[serde(rename = "login")]
-    pub login: &'a str,
-    #[serde(rename = "username")]
-    pub username: &'a str,
-    #[serde(rename = "password")]
-    pub password: &'a str,
-}
